Update patch set 1
Patch Set 1:

Yeah 18dec91 is the one that fixed building for me when I was trying to bring up hlte. Looking at the commit it makes sense to me that it should be needed, because otherwise the kernel isn't an explicit dependency for keymaster on some devices (if I understand the build system correctly).

Patch-set: 1